js separado
