+++ app/views/layouts/user.php
@@ -1,3 +1,4 @@
+
 <?php
 function isActive($route, $currentUrl)
 {
