--- src/logout.php
@@ -1,7 +1,0 @@
-<?php
-session_start();
-session_unset();  // Limpia todas las variables de sesión
-session_destroy(); // Destruye la sesión
-header("Location: login.php"); // Redirige al login
-exit();
-?>
